+++ examples/javaScript/sell/account/getSalesTax.js
@@ -4,10 +4,9 @@ var clientId = process.env.EBAY_CLIENT_ID || 'YOUR_KEY';
 var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 
 var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
-
 var utils = require('../../../javaScript/utils');
 
-var catalogRequest = async function () {
+var accountRequest = async function () {
     var userToken = utils.USER_TOKEN;
     eBay.setUserToken(userToken);
 
@@ -22,4 +21,4 @@ var catalogRequest = async function () {
     }
 };
 
-catalogRequest();
+accountRequest();