+++ examples/javaScript/sell/inventory/getInventoryItem.js
@@ -7,7 +7,6 @@ var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
 var utils = require('../../../javaScript/utils');
 
 var inventoryRequest = async function () {
-
     var userToken = utils.USER_TOKEN;
     eBay.setUserToken(userToken);