+++ examples/javaScript/sell/offer/getListingFees.js
@@ -6,15 +6,20 @@ var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
 var utils = require('../../../javaScript/utils');
 
-var accountRequest = async function () {
+var offerRequest = async function () {
     var userToken = utils.USER_TOKEN;
     eBay.setUserToken(userToken);
 
+    var data = {
+        offers: [
+            {
+                'offerId': '6360337010'
+            }
+        ]
+    };
 
-    var countryCode = 'US';
-    var jurisdictionId = 'IN';
     try {
-        var response = await eBay.account.deleteSalesTax(countryCode, jurisdictionId);
+        var response = await eBay.offer.getListingFees(data);
         console.log('response', response);
     } catch (error) {
         console.log('error ', error);
@@ -22,4 +27,4 @@ var accountRequest = async function () {
     }
 };
 
-accountRequest();
+offerRequest();